examples/simple-screenshot/requirements.txt
flask==3.0.0
selenium==4.15.2
pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0